Make ClockVolume Optional
ClockVolume put a dependency on the render pipeline core library, and for games to use a render pipeline in order to work with time. This update makes the clock volume completely optional: handy if you want it, but not required.
