+++ Runtime/CoreRPLibrary.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f151c58a6a6425b418a360bf0d23d572
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
