+++ Runtime/CoreRPLibrary/CaseyHofland.UnityClock.RenderPipelines.Core.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c0841c202898cf24dbf9617181cc7238
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
